+++ src/components/TaskItem.tsx
@@ -7,6 +7,14 @@ import {
   Typography,
   Paper,
   useTheme,
+  Dialog,
+  DialogTitle,
+  DialogContent,
+  DialogContentText,
+  DialogActions,
+  Button,
+  Box,
+  Stack,
 } from "@mui/material";
 // MUI Icons
 import AccessTimeIcon from "@mui/icons-material/AccessTime";
@@ -17,18 +25,13 @@ import useTaskStore from "../store/taskStore";
 // Types
 import taskType from "../types/taskType";
 // React Router
-import { useNavigate } from "react-router";
+import { Link, useNavigate } from "react-router";
 // Utils
 import { EDIT_TASK_ROUTE } from "../utils/fullRoutes";
 // Other external packages
 import { useLongPress } from "use-long-press";
-import Menu from "./Menu";
 import { useMemo, useState } from "react";
-
-import MoreVertIcon from "@mui/icons-material/MoreVert";
-import { grey } from "@mui/material/colors";
-
-import Dialog from "./Dialog";
+import { grey, green, orange, red } from "@mui/material/colors";
 
 type props = {
   task: taskType;
@@ -40,7 +43,7 @@ const TaskItem = ({ task }: props) => {
 
   const theme = useTheme();
   const toggleComplete = useTaskStore((state) => state.toggleComplete);
-  const removeTask = useTaskStore((state) => state.removeTask);
+  const deleteTask = useTaskStore((state) => state.deleteTask);
   const navigate = useNavigate();
   const bind = useLongPress((taskId) => {
     console.log("USE LONG PRESS");
@@ -48,41 +51,102 @@ const TaskItem = ({ task }: props) => {
     navigate(`${EDIT_TASK_ROUTE}/${taskId}`);
   });
 
-  const menuItems = useMemo(
-    () => [
-      {
-        name: "Edit",
-        onClick: () => {
-          navigate(`${EDIT_TASK_ROUTE}/${task.id}`);
-        },
-      },
-      {
-        name: "Delete",
-        onClick: () => {
-          removeTask(task.id);
-        },
-      },
-    ],
-    [navigate, task.id]
-  );
+  // The colour of the left border represents the priority,
+  // or no colour if no priority. The colour shade depends on
+  // whether the application is on light or dark mode. Return the
+  // appropriate colour based on these factors.
+  const getLeftBorderColour = () => {
+    // If light mode
+    if (theme.palette.mode === "light") {
+      return task.priority === "low"
+        ? green[200]
+        : task.priority === "medium"
+        ? orange[200]
+        : task.priority === "high"
+        ? red[200]
+        : "transparent";
+    }
+
+    // If dark mode
+    return task.priority === "low"
+      ? green[800]
+      : task.priority === "medium"
+      ? orange[800]
+      : task.priority === "high"
+      ? red[800]
+      : "transparent";
+  };
+
+  // const menuItems = useMemo(
+  //   () => [
+  //     {
+  //       name: "Edit",
+  //       onClick: () => {
+  //         navigate(`${EDIT_TASK_ROUTE}/${task.id}`);
+  //       },
+  //     },
+  //     {
+  //       name: "Delete",
+  //       onClick: () => {
+  //         removeTask(task.id);
+  //       },
+  //     },
+  //   ],
+  //   [navigate, task.id]
+  // );
 
   return (
     <>
       <Paper
+        draggable
         sx={{
+          height: "50px",
           backgroundColor:
-            theme.palette.mode === "light" ? grey[50] : grey[900],
+            theme.palette.mode === "light" ? grey[100] : grey[900],
           display: "flex",
           justifyContent: "space-between",
           alignItems: "center",
+          borderLeftWidth: "5px",
+          borderLeftStyle: "solid",
+          borderLeftColor: getLeftBorderColour,
         }}
       >
-        <Typography>{task.name}</Typography>
-        <IconButton>
-          <DeleteIcon fontSize="small" onClick={() => setOpen(true)} />
-        </IconButton>
+        <Stack>
+          <Typography>{task.name}</Typography>
+          <Typography sx={{ fontSize: "0.7rem", fontWeight: "300" }}>
+            {task.description}
+          </Typography>
+        </Stack>
+        <Stack direction="row">
+          <IconButton onClick={() => navigate(`${EDIT_TASK_ROUTE}/${task.id}`)}>
+            <EditIcon fontSize="small" />
+          </IconButton>
+          <IconButton onClick={() => setOpen(true)}>
+            <DeleteIcon fontSize="small" />
+          </IconButton>
+        </Stack>
       </Paper>
-      <Dialog />
+      <Dialog open={open} onClose={() => setOpen(false)}>
+        <DialogTitle>Delete Task</DialogTitle>
+        <DialogContent>
+          <DialogContentText>
+            Are you sure you want to delete the task '{task.name}'?
+          </DialogContentText>
+        </DialogContent>
+        <DialogActions>
+          <Button
+            onClick={() => {
+              setOpen(false);
+              deleteTask(task.id);
+            }}
+          >
+            Yes
+          </Button>
+          <Button onClick={() => setOpen(false)} autoFocus>
+            No
+          </Button>
+        </DialogActions>
+      </Dialog>
     </>
     // <Card
     //   variant="outlined"
@@ -93,13 +157,13 @@ const TaskItem = ({ task }: props) => {
     //     display: "flex",
     //     justifyContent: "space-between",
     //     borderLeft:
-    //       task.priority === "low"
-    //         ? "2px solid green"
-    //         : task.priority === "medium"
-    //         ? "2px solid orange"
-    //         : task.priority === "high"
-    //         ? "2px solid red"
-    //         : "",
+    // task.priority === "low"
+    //   ? "2px solid green"
+    //   : task.priority === "medium"
+    //   ? "2px solid orange"
+    //   : task.priority === "high"
+    //   ? "2px solid red"
+    //   : "",
     //   }}
     // >
     //   <CardContent